fix: premium slot count always 0 due to BIGINT-as-string concat
`getLaunchAvailability` used `count(*)` without a `::int` cast.
PostgreSQL returns BIGINT, which pg/Drizzle serialises to a JS
string despite the `sql<number>` type hint. The formula
  `PREMIUM_DAILY_LIMIT - (premium + free + badge)`
then triggered string concatenation:
  `10 - ("0" + "5" + "0") = 10 - "050" = -40 → max(0, -40) = 0`

Every date rendered as "0 premium slots available" and disabled,
blocking all Boost-tier submissions. The free path silently
worked because `freeSlots = max(0, 5 - "5") = 0` happens to be
the correct value when the free queue is full (string -> number
coercion works for single-term subtraction).

Same cast pattern is already used in `scheduleLaunch` lower in
the file — this aligns `getLaunchAvailability` with it.

+++ app/actions/launch.ts
@@ -49,10 +49,15 @@ export async function getLaunchAvailability(date: string): Promise<LaunchAvailab
   // Obtenir le nombre de lancements déjà programmés pour cette date
   const scheduledLaunches = await db
     .select({
-      freeCount: sql<number>`count(*) filter (where ${projectTable.launchType} = ${launchType.FREE})`,
-      badgeCount: sql<number>`count(*) filter (where ${projectTable.launchType} = 'free_with_badge')`,
-      premiumCount: sql<number>`count(*) filter (where ${projectTable.launchType} = ${launchType.PREMIUM})`,
-      totalCount: sql<number>`count(*)`,
+      // `::int` casts the BIGINT count to int4 so pg/Drizzle hands
+      // back a JS number, not a string. Without it, the arithmetic
+      // below silently turns into string concatenation:
+      //   `10 - ("0" + "5" + "0")` → `10 - "050"` → -40 → max(0, -40) = 0
+      // and every date renders as "0 premium slots available".
+      freeCount: sql<number>`count(*) filter (where ${projectTable.launchType} = ${launchType.FREE})::int`,
+      badgeCount: sql<number>`count(*) filter (where ${projectTable.launchType} = 'free_with_badge')::int`,
+      premiumCount: sql<number>`count(*) filter (where ${projectTable.launchType} = ${launchType.PREMIUM})::int`,
+      totalCount: sql<number>`count(*)::int`,
     })
     .from(projectTable)
     .where(
